refactor:"PostItem" divided into "PostItem" and "PostItemDialog"
