+++ src/components/molecules/PostItem.tsx
@@ -1,29 +1,17 @@
 import React, { useCallback, memo, ChangeEvent, useState } from "react";
-import { useRecoilValue } from "recoil";
-import {
-  isFollowedState,
-  userState,
-  viewedUserState,
-} from "../../store/atoms/userAtom";
 import { useAllParks } from "../../hooks/useAllParks";
-import { FollowButton } from "../atoms/button/FollowButton";
+import { PostItemDialog } from "./PostItemDialog";
 import format from "date-fns/format";
 import ja from "date-fns/locale/ja";
 import {
-  Avatar,
-  Box,
-  Button,
   Card,
   CardContent,
   CardMedia,
   Checkbox,
-  Dialog,
-  DialogContent,
   FormControlLabel,
   Typography,
   styled,
 } from "@mui/material";
-import { ArrowBackIos, ArrowForwardIos } from "@mui/icons-material";
 import { Post } from "../../types/images";
 
 type Props = {
@@ -66,9 +54,6 @@ export const PostItem = memo((props: Props) => {
   } = props;
 
   const { parks } = useAllParks();
-  const user = useRecoilValue(userState);
-  const viewedUser = useRecoilValue(viewedUserState);
-  const isFollowed = useRecoilValue(isFollowedState);
 
   const handleChange = useCallback(
     (e: ChangeEvent<HTMLInputElement>) => {
@@ -166,141 +151,20 @@ export const PostItem = memo((props: Props) => {
           />
           {currentPost && (
             <>
-              <Dialog
-                open={isDialogVisible && imageOpen}
-                onClose={handleClickImageClose}
-                fullWidth={true}
-                maxWidth={"xl"}
-                sx={{
-                  backgroundColor: "rgba(0,0,0,0.9)",
-                  display: "flex",
-                  flexDirection: "column",
-                  justifyContent: "center",
-                  alignItems: "center",
-                }}
-                onClick={(e) => e.stopPropagation()}
-              >
-                <DialogContent
-                  sx={{
-                    display: "flex",
-                    justifyContent: "center",
-                    alignItems: "flex-start",
-                    overflow: "auto",
-                    margin: 0,
-                    padding: 0,
-                    width: "100%",
-                    height: "100%",
-                  }}
-                >
-                  <Box component="span">
-                    <img
-                      src={currentPost.image}
-                      alt="currentPostImage"
-                      style={{
-                        objectFit: "contain",
-                        height: imageSize.height,
-                        width: imageSize.width,
-                        display: "block",
-                      }}
-                    />
-                  </Box>
-                  <Box
-                    component="span"
-                    display="flex"
-                    flexDirection="column"
-                    alignItems="flex-start"
-                    gap={1.5}
-                    p={2}
-                    width="250px"
-                    overflow="auto"
-                  >
-                    <Box display="flex" flexDirection="row">
-                      <Avatar
-                        src={
-                          viewedUser?.avatar instanceof File
-                            ? URL.createObjectURL(viewedUser.avatar)
-                            : viewedUser?.avatar || ""
-                        }
-                      />
-                      <Typography
-                        variant="h6"
-                        paddingLeft="10px"
-                        paddingRight="10px"
-                      >
-                        {viewedUser?.nickname}
-                      </Typography>
-                      {user?.id !== viewedUser?.id && isFollowed === false && (
-                        <FollowButton
-                          handleFollowButtonClick={() =>
-                            handleFollowButtonClick(numUserId, isFollowed)
-                          }
-                          isFollowed={isFollowed}
-                        />
-                      )}
-                    </Box>
-                    <Typography variant="body1">
-                      昆虫名:{" "}
-                      {currentPost.insect_name
-                        ? `${currentPost.insect_name}(${currentPost.insect_sex})`
-                        : "\u00a0"}
-                    </Typography>
-                    <Typography variant="body1">
-                      撮影場所:{" "}
-                      {currentPost.park_id !== null &&
-                      parks[currentPost.park_id - 1]?.name
-                        ? parks[currentPost.park_id - 1]?.name
-                        : currentPost.city_name}
-                    </Typography>
-                    <Typography variant="body1">
-                      撮影日時:{" "}
-                      {createdTime(currentPost)
-                        ? createdTime(currentPost)
-                        : "\u00a0"}
-                    </Typography>
-                  </Box>
-                </DialogContent>
-                {currentImageIndex !== 0 && (
-                  <Button
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      handlePrevImageClick();
-                    }}
-                    sx={{
-                      position: "fixed",
-                      left: "1%",
-                      top: "50%",
-                      transform: "translateY(-50%)",
-                      zIndex: 1500,
-                      color: "darkgray",
-                      borderRadius: "10%",
-                      "&:hover": {
-                        backgroundColor: "rgba(0,0,0,0.5)",
-                      },
-                    }}
-                  >
-                    <ArrowBackIos />
-                  </Button>
-                )}
-                {currentImageIndex !== maxIndex && (
-                  <Button
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      handleNextImageClick();
-                    }}
-                    sx={{
-                      position: "fixed",
-                      right: "1%",
-                      top: "50%",
-                      transform: "translateY(-50%)",
-                      zIndex: 1500,
-                      color: "darkgray",
-                      borderRadius: "10%",
-                    }}
-                  >
-                    <ArrowForwardIos />
-                  </Button>
-                )}
-              </Dialog>
+              <PostItemDialog
+                numUserId={numUserId}
+                currentPost={currentPost}
+                currentImageIndex={currentImageIndex}
+                maxIndex={maxIndex}
+                createdTime={createdTime}
+                isDialogVisible={isDialogVisible}
+                imageOpen={imageOpen}
+                imageSize={imageSize}
+                handleClickImageClose={handleClickImageClose}
+                handlePrevImageClick={handlePrevImageClick}
+                handleNextImageClick={handleNextImageClick}
+                handleFollowButtonClick={handleFollowButtonClick}
+              />
             </>
           )}
         </SquareCard>
